Move GuidanceModal in own component
